have simulate evaluate its arrays
previously `simulate` would just build the Acc expression and it was the renderer which called `run` to evaluate it, but this would lead to the expression growing exponentially if render is not called after each simulation step.

+++ src/Config.hs
@@ -110,7 +110,7 @@ defaultConfig =  Config
       &= name "s"
       &= name "scale"
       &= help "feature size of visualisation"
-      &= typ "FLOAT"
+      &= typ "INT"
 
   , displayFramerate = 10
       &= explicit
